refactor: move HTML style handling from UIWidget.cpp to UIWidgetHtml.cpp
