+++ src/framework/pch.h
@@ -48,6 +48,7 @@
 #include <unordered_map>
 #include <unordered_set>
 #include <vector>
+#include <ranges>
 
 #include <parallel_hashmap/btree.h>
 #include <parallel_hashmap/phmap.h>
